+++ crates/db_schema/src/source/person.rs
@@ -56,12 +56,12 @@ pub struct Person {
   /// Whether the person is a bot account.
   pub bot_account: bool,
   pub instance_id: InstanceId,
-  pub post_count: i64,
+  pub post_count: i32,
   #[serde(skip)]
-  pub post_score: i64,
-  pub comment_count: i64,
+  pub post_score: i32,
+  pub comment_count: i32,
   #[serde(skip)]
-  pub comment_score: i64,
+  pub comment_score: i32,
 }
 
 #[derive(Clone, derive_new::new)]
